+++ forge/src/main/java/net/mehvahdjukaar/hauntedharvest/forge/HauntedHarvestForgeClient.java
@@ -10,6 +10,7 @@ import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.item.Item;
 import net.minecraftforge.client.event.RenderNameTagEvent;
 import net.minecraftforge.common.MinecraftForge;
+import net.minecraftforge.eventbus.api.Event;
 import net.minecraftforge.eventbus.api.IEventBus;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.ModList;
@@ -27,16 +28,15 @@ public class HauntedHarvestForgeClient {
             }
         });
 
-        MinecraftForge.EVENT_BUS.register(HauntedHarvestForgeClient.class);
+        MinecraftForge.EVENT_BUS.addListener(HauntedHarvestForgeClient::onRenderNameTag);
 
     }
 
-    @SubscribeEvent
     public static void onRenderNameTag(RenderNameTagEvent event) {
         if (CommonConfigs.PAPER_BAG_NAME_TAG.get() && event.getEntity() instanceof LivingEntity le) {
             Item slot = le.getItemBySlot(EquipmentSlot.HEAD).getItem();
             if (slot == ModRegistry.PAPER_BAG_ITEM.get()) {
-                event.setCanceled(true);
+                event.setResult(Event.Result.DENY);
             }
         }
     }