+++ changelog.md
@@ -1,3 +1,5 @@
 - paper bag can hide name tags
 - paper bags can hide from endermen
+- paper bags now have an overlay texture
 - fixed paper bag head inner model
+- fixed pumpkin overlay disappearing when reloading textures